little refactor
