+++ packages/netlify-cms-media-library-uploadcare/src/components/MediaLibraryCard.js
