+++ packages/netlify-cms-media-library-uploadcare/src/components/MediaLibrarySearch.js
